+++ changelog.md
@@ -1,6 +1,7 @@
 # Changelog for Bolt Forms
 
 * 2.x.x (2015-xx-xx)
+ * Added: Form specific template 
  * Fixed: Sending of CC and bCC recipients
  * Break: Parameters for Processor::process() changed to allow better API use of custom form values